+++ app/src/main/java/com/distrigo/app/data/local/entity/ProductEntity.kt
@@ -11,7 +11,7 @@ data class ProductEntity(
     val barcode: String?,
     val selling_price: Double,
     val purchase_price: Double,
-    val stock: Int,
+    val stock: Double,
     val min_stock: Int,
     val unit_type: String,
     val packages: Int,
@@ -30,5 +30,4 @@ data class ProductEntity(
     val supplier_name: String?,
     val supplier_id: Int?,
 
-    val camion_stock: Int = 0
-)
+    val camion_stock: Double = 0.0)
